+++ dot_oh-my-zsh/custom/plugins/zsh-syntax-highlighting/tests/executable_test-highlighting.zsh
@@ -144,7 +144,8 @@ run_test_internal() {
     }
 
     # Check the data declares $PREBUFFER or $BUFFER.
-    [[ -z $PREBUFFER && -z $BUFFER ]] && { echo >&2 "Bail out! On ${(qq)1}: Either 'PREBUFFER' or 'BUFFER' must be declared and non-blank"; return ${RETURN:=1}; }
+    [[ -z $PREBUFFER && -z $BUFFER ]] && { echo >&2 "Bail out! On ${(qq)ARG}: Either 'PREBUFFER' or 'BUFFER' must be declared and non-blank"; return ${RETURN:=1}; }
+    [[ $PREBUFFER == (''|*$'\n') ]] || { echo >&2 "Bail out! On ${(qq)ARG}: PREBUFFER=${(qqqq)PREBUFFER} doesn't end with a newline"; return ${RETURN:=1}; }
 
     # Set sane defaults for ZLE variables
     : ${CURSOR=$#BUFFER} ${PENDING=0} ${WIDGET=z-sy-h-test-harness-test-widget}
@@ -153,6 +154,8 @@ run_test_internal() {
     _zsh_highlight
   }; [[ -z $RETURN ]] || return $RETURN
   unset ARG
+
+  integer print_expected_and_actual=0
 
   if (( unsorted )); then
     region_highlight=("${(@n)region_highlight}")
@@ -177,6 +180,7 @@ run_test_internal() {
     if ! (( $+region_highlight[i] )); then
       print -r -- "not ok $i - unmatched expectation ($exp_start $exp_end $expected_highlight_zone[3])" \
          "${skip_mismatch:+"# TODO ${(qqq)skip_mismatch}"}"
+      if [[ -z $skip_mismatch ]]; then (( ++print_expected_and_actual )); fi
       continue
     fi
     local -a highlight_zone; highlight_zone=( ${(z)region_highlight[i]} )
@@ -189,6 +193,7 @@ run_test_internal() {
       [[ $highlight_zone[3] != $expected_highlight_zone[3] ]]
     then
       print -r -- "not ok $i - $desc - expected ($exp_start $exp_end ${(qqq)expected_highlight_zone[3]}), observed ($start $end ${(qqq)highlight_zone[3]}). $todo"
+      if [[ -z $todo ]]; then (( ++print_expected_and_actual )); fi
     else
       print -r -- "ok $i - $desc${todo:+ - }$todo"
     fi
@@ -221,7 +226,10 @@ run_test_internal() {
       details+="«$(typeset_p expected_region_highlight)» «$(typeset_p region_highlight)»"
       tap_escape $details; details=$REPLY
       print -r -- "not ok $i - cardinality check - $details${todo:+ - }$todo"
-
+      if [[ -z $todo ]]; then (( ++print_expected_and_actual )); fi
+    fi
+  fi
+  if (( print_expected_and_actual )); then
       () {
         local -a left_column right_column
         left_column=( "expected_region_highlight" "${(qq)expected_region_highlight[@]}" )
@@ -234,7 +242,6 @@ run_test_internal() {
           | if type column >/dev/null; then column -t -s $'\t'; else cat; fi \
           | sed 's/^/# /'
       }
-    fi
   fi
 }